packages/cleafy-plugin/android/src/main/kotlin/com/cleafy/cleafyPlugin/methods/AppSessionIdFlutterMethod.kt
package com.cleafy.cleafyPlugin.methods

import com.cleafy.cleafyPlugin.interop.CleafyProxy
import io.flutter.embedding.engine.plugins.FlutterPlugin
import io.flutter.plugin.common.MethodCall
import io.flutter.plugin.common.MethodChannel

internal class AppSessionIdFlutterMethod(private val cleafy: CleafyProxy) : FlutterMethod {

    override fun invoke(
        flutterPluginBinding: FlutterPlugin.FlutterPluginBinding,
        call: MethodCall,
        result: MethodChannel.Result
    ) {
        val sessionId = call.arguments<String>() ?: run {
            result.error("", "App session ID not found", null)
            return
        }
        cleafy.setAppSessionId(sessionId)
        result.success("New app session ID set!")
    }
}
